backend/requirements.txt
Flask==3.0.0
pandas==2.1.3
pymongo==4.3.3
html2image==2.0.4.3
numpy==1.23.5
pandas==2.1.3
matplotlib==3.8.2
calmap==0.0.11
python-dotenv==1.0.0